services/moderation-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
aio-pika==9.3.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
slowapi==0.1.9
python-dotenv==1.0.0
pyjwt==2.8.0
redis==5.0.1
prometheus-fastapi-instrumentator==6.1.0

